Extract input fetching to separate file
